requirements.txt
fastapi
pydantic
python-dotenv
uvicorn
starlette
python-multipart
jinja2
openai
langchain==0.3.4
unstructured[xlsx,docx,pptx,xlsx]==0.16.0
langchain-community==0.3.3
langchain-openai==0.2.3
langchain-unstructured
markdown==3.5.1
redis==5.1.1
tiktoken
pyjwt
azure-storage-blob
diskcache
azure-ai-translation-document
azure-ai-translation-text
azure-cognitiveservices-speech
srt
pydub
aiofiles